python-service/requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.1
Flask-Migrate==3.1.0
python-dotenv
flask_cors